Create State.jsx
In React, the useState Hook lets you add state to functional components. It gives you two things: the current state and a function to change it. You start by giving it an initial value, and whenever you use the function to update the state, React gives you the new value.